% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_engine_examdesign.R
\name{engine.analyze.class.examdesign}
\alias{engine.analyze.class.examdesign}
\title{Function that breaks latex file of exam class examdesign into a dataframe (internal use)}
\usage{
engine.analyze.class.examdesign(my.text)
}
\arguments{
\item{my.text}{Text of latex file written in class examdesign (UTF-8)}
}
\value{
A dataframe with several information (see rte.analyze.tex.file)
}
\description{
Function that breaks latex file of exam class examdesign into a dataframe (internal use)
}
\examples{
f.in <- system.file("extdata", "MyRandomTest_examdesign.tex", package = "RndTexExams")
my.text <- stringi::stri_read_lines(f.in)
Encoding(my.text) <- 'UTF-8'

out <- engine.analyze.class.examdesign(my.text)

}
