% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_Utils.R
\name{rte.check.latex.flavor}
\alias{rte.check.latex.flavor}
\title{Function to check the distribution of LaTeX}
\usage{
rte.check.latex.flavor()
}
\value{
The flavor of latex instlation (e.g. miktex, texlive)
}
\description{
Function to check the distribution of LaTeX
}
\examples{
rte.check.latex.flavor()

}
