% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_Utils.R
\name{rte.check.my.os}
\alias{rte.check.my.os}
\title{Function to check operating system of user}
\usage{
rte.check.my.os()
}
\value{
A string with the name of the operating system (e.g. Windows)
}
\description{
Function to check operating system of user
}
\examples{
rte.check.my.os()

}

