% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_Utils.R
\name{rte.check.pdflatex}
\alias{rte.check.pdflatex}
\title{Function to check if system has pdflatex.exe available}
\usage{
rte.check.pdflatex()
}
\value{
TRUE if the pdflatex is available, FALSE if not
}
\description{
Function to check if system has pdflatex.exe available
}
\examples{
rte.check.pdflatex()

}
