% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_Utils.R
\name{rte.check.system.command}
\alias{rte.check.system.command}
\title{Function to check if system command is available}
\usage{
rte.check.system.command(mycommand)
}
\arguments{
\item{mycommand}{A system command (e.g. pdflatex inputfile.tex)}
}
\value{
A flag, TRUE if the command is available  or FALSE if not)
}
\description{
Function to check if system command is available
}
\examples{
rte.check.system.command('pdflatex')

}

