% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_Utils.R
\name{rte.compile.latex}
\alias{rte.compile.latex}
\title{Function to compile a LaTeX file}
\usage{
rte.compile.latex(f.in, pdf.dir.out = "PdfOut")
}
\arguments{
\item{f.in}{The location and name of latex file}

\item{pdf.dir.out}{The name of the folder for the output pdf}
}
\value{
A flag, TRUE if the latex compilation was a sucess and FALSE if not
}
\description{
This function will first check for the user's OS and then use the proper
command for pdflatex compilation
}
\examples{
f.in <- system.file("extdata", "MyRandomTest.tex", package = "RndTexExams")
pdf.dir.out <- 'PdfOut'

\dontrun{rte.compile.latex <- function(f.in = f.in,
                             pdf.dir.out =  pdf.dir.out) }

}

