% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_Utils.R
\name{rte.get.classes.def}
\alias{rte.get.classes.def}
\title{Function that returns parameters of latex classes (internal use)}
\usage{
rte.get.classes.def(exam.class)
}
\arguments{
\item{exam.class}{The class of the exam (exam, examdesign)}
}
\value{
A list with parameters of the class
}
\description{
This function outputs the main elements of each latex class such as question/choices identifiers and environmental switches.
}
\examples{
l.out <- rte.get.classes.def(exam.class = 'exam')
print(l.out)
}
