% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_Utils.R
\name{rte.get.n.cases}
\alias{rte.get.n.cases}
\title{Function to get number of cases (textual switches) based on question text}
\usage{
rte.get.n.cases(str.in)
}
\arguments{
\item{str.in}{The full text of the question}
}
\value{
n.cases Number of cases (textual switches in string)
}
\description{
Function to get number of cases (textual switches) based on question text
}
\examples{

my.question <- 'My questions is ... @{ver1}|{ver2}@' ## two cases

n.cases <- rte.get.n.cases(my.question)

}

