% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte_The_Randomizer.R
\name{rte.The.Randomizer}
\alias{rte.The.Randomizer}
\title{Function to randomize a question in a dataframe (internal use)}
\usage{
rte.The.Randomizer(q.text, q.answers, case.now, my.rdn.idx.answers)
}
\arguments{
\item{q.text}{Main text of question (character)}

\item{q.answers}{The answers of the questions as an atomic vector (each alternative as an item)}

\item{case.now}{The random case to build the string (only used for textual switches)}

\item{my.rdn.idx.answers}{The random index of the questions}
}
\value{
A list with the text of the full questions, among other things
}
\description{
Function to randomize a question in a dataframe (internal use)
}
\examples{
q.text <- '\\\\question Whats my name \\\\n'
q.answers <- c('\\\\choice Mario', '\\\\choice Roberto','\\\\choice Marcelo')

case.now <- 1
my.rdn.idx.answers <- sample(seq(length(q.answers)))

l.out <- rte.The.Randomizer(q.text,q.answers, case.now, my.rdn.idx.answers)
}

